Added fix for missing ISynchronizedStorage configuration
