+++ src/NServiceBus.AcceptanceTests/EndpointTemplates/ConfigureExtensions.cs
@@ -4,6 +4,7 @@
     using System.Threading.Tasks;
     using NServiceBus.AcceptanceTesting.Support;
     using NServiceBus.AcceptanceTests.ScenarioDescriptors;
+    using ObjectBuilder;
 
     public static class ConfigureExtensions
     {
@@ -27,6 +28,47 @@
             }
 
             return ConfigureTestExecution(TestDependencyType.Persistence, config, settings, endpointName);
+        }
+
+        public static void DefineBuilder(this EndpointConfiguration config, RunSettings settings)
+        {
+            Type builderType;
+            if (!settings.TryGet("Builder", out builderType))
+            {
+                var builderDescriptor = Builders.Default;
+
+                if (builderDescriptor == null)
+                {
+                    return; //go with the default builder
+                }
+
+                settings.Merge(builderDescriptor.Settings);
+            }
+
+            builderType = settings.Get<Type>("Builder");
+
+            var typeName = "Configure" + builderType.Name;
+
+            var configurerType = Type.GetType(typeName, false);
+
+            if (configurerType != null)
+            {
+                var configurer = Activator.CreateInstance(configurerType);
+
+                dynamic dc = configurer;
+
+                dc.Configure(config);
+            }
+
+            config.UseContainer(builderType);
+        }
+
+        public static void RegisterComponentsAndInheritanceHierarchy(this EndpointConfiguration builder, RunDescriptor runDescriptor)
+        {
+            builder.RegisterComponents(r =>
+            {
+                RegisterInheritanceHierarchyOfContextOnContainer(runDescriptor, r);
+            });
         }
 
         static async Task ConfigureTestExecution(TestDependencyType type, EndpointConfiguration config, RunSettings settings, string endpointName)
@@ -63,37 +105,14 @@
             cleaners.Add(configurer);
         }
 
-        public static void DefineBuilder(this EndpointConfiguration config, RunSettings settings)
+        static void RegisterInheritanceHierarchyOfContextOnContainer(RunDescriptor runDescriptor, IConfigureComponents r)
         {
-            Type builderType;
-            if (!settings.TryGet("Builder", out builderType))
+            var type = runDescriptor.ScenarioContext.GetType();
+            while (type != typeof(object))
             {
-                var builderDescriptor = Builders.Default;
-
-                if (builderDescriptor == null)
-                {
-                    return; //go with the default builder
-                }
-
-                settings.Merge(builderDescriptor.Settings);
+                r.RegisterSingleton(type, runDescriptor.ScenarioContext);
+                type = type.BaseType;
             }
-
-            builderType = settings.Get<Type>("Builder");
-
-            var typeName = "Configure" + builderType.Name;
-
-            var configurerType = Type.GetType(typeName, false);
-
-            if (configurerType != null)
-            {
-                var configurer = Activator.CreateInstance(configurerType);
-
-                dynamic dc = configurer;
-
-                dc.Configure(config);
-            }
-
-            config.UseContainer(builderType);
         }
 
         enum TestDependencyType
